Intermediate Commit 5

+++ src/analyzer/Analyzer.cc
@@ -307,7 +307,7 @@ void Analyzer::ForwardPacket(int len, const u_char* data, bool is_orig,
 	AppendNewChildren();
 	}
 
-void Analyzer::ForwardStream(int len, const u_char* data, bool is_orig)
+void Analyzer::ForwardStream(uint64_t len, const u_char* data, bool is_orig)
 	{
 	if ( output_handler )
 		output_handler->DeliverStream(len, data, is_orig);
@@ -628,7 +628,7 @@ SupportAnalyzer* Analyzer::FirstSupportAnalyzer(bool orig)
 	return sa->Sibling(true);
 	}
 
-void Analyzer::DeliverPacket(int len, const u_char* data, bool is_orig,
+void Analyzer::DeliverPacket(uint64_t len, const u_char* data, bool is_orig,
 				uint64_t seq, const IP_Hdr* ip, int caplen)
 	{
 	DBG_LOG(DBG_ANALYZER, "%s DeliverPacket(%d, %s, %" PRIu64", %p, %d) [%s%s]",
@@ -636,14 +636,14 @@ void Analyzer::DeliverPacket(int len, const u_char* data, bool is_orig,
 			fmt_bytes((const char*) data, min(40, len)), len > 40 ? "..." : "");
 	}
 
-void Analyzer::DeliverStream(int len, const u_char* data, bool is_orig)
+void Analyzer::DeliverStream(uint64_t len, const u_char* data, bool is_orig)
 	{
 	DBG_LOG(DBG_ANALYZER, "%s DeliverStream(%d, %s) [%s%s]",
 			fmt_analyzer(this).c_str(), len, is_orig ? "T" : "F",
 			fmt_bytes((const char*) data, min(40, len)), len > 40 ? "..." : "");
 	}
 
-void Analyzer::Undelivered(uint64_t seq, int len, bool is_orig)
+void Analyzer::Undelivered(uint64_t seq, uint64_t len, bool is_orig)
 	{
 	DBG_LOG(DBG_ANALYZER, "%s Undelivered(%" PRIu64", %d, %s)",
 			fmt_analyzer(this).c_str(), seq, len, is_orig ? "T" : "F");
@@ -856,7 +856,7 @@ void SupportAnalyzer::ForwardPacket(int len, const u_char* data, bool is_orig,
 		Parent()->DeliverPacket(len, data, is_orig, seq, ip, caplen);
 	}
 
-void SupportAnalyzer::ForwardStream(int len, const u_char* data, bool is_orig)
+void SupportAnalyzer::ForwardStream(uint64_t len, const u_char* data, bool is_orig)
 	{
 	// We do not call parent's method, as we're replacing the functionality.
