+++ src/analyzer/protocol/bittorrent/BitTorrent.h
@@ -15,8 +15,8 @@ public:
 	~BitTorrent_Analyzer() override;
 
 	void Done() override;
-	void DeliverStream(int len, const u_char* data, bool orig) override;
-	void Undelivered(uint64_t seq, int len, bool orig) override;
+	void DeliverStream(uint64_t len, const u_char* data, bool orig) override;
+	void Undelivered(uint64_t seq, uint64_t len, bool orig) override;
 	void EndpointEOF(bool is_orig) override;
 
 	static analyzer::Analyzer* Instantiate(Connection* conn)
