+++ src/analyzer/protocol/conn-size/ConnSize.cc
@@ -100,7 +100,7 @@ void ConnSize_Analyzer::CheckThresholds(bool is_orig)
 		}
 	}
 
-void ConnSize_Analyzer::DeliverPacket(int len, const u_char* data, bool is_orig, uint64_t seq, const IP_Hdr* ip, int caplen)
+void ConnSize_Analyzer::DeliverPacket(uint64_t len, const u_char* data, bool is_orig, uint64_t seq, const IP_Hdr* ip, int caplen)
 	{
 	Analyzer::DeliverPacket(len, data, is_orig, seq, ip, caplen);
