+++ src/analyzer/protocol/conn-size/ConnSize.h
@@ -31,7 +31,7 @@ public:
 		{ return new ConnSize_Analyzer(conn); }
 
 protected:
-	void DeliverPacket(int len, const u_char* data, bool is_orig,
+	void DeliverPacket(uint64_t len, const u_char* data, bool is_orig,
 					   uint64_t seq, const IP_Hdr* ip, int caplen) override;
 	void CheckThresholds(bool is_orig);
