+++ src/analyzer/protocol/dce-rpc/DCE_RPC.h
@@ -18,8 +18,8 @@ public:
 	~DCE_RPC_Analyzer() override;
 
 	void Done() override;
-	void DeliverStream(int len, const u_char* data, bool orig) override;
-	void Undelivered(uint64_t seq, int len, bool orig) override;
+	void DeliverStream(uint64_t len, const u_char* data, bool orig) override;
+	void Undelivered(uint64_t seq, uint64_t len, bool orig) override;
 	void EndpointEOF(bool is_orig) override;
 
 	bool SetFileID(uint64_t fid_in)
