+++ src/analyzer/protocol/dhcp/DHCP.cc
@@ -21,7 +21,7 @@ void DHCP_Analyzer::Done()
 	Analyzer::Done();
 	}
 
-void DHCP_Analyzer::DeliverPacket(int len, const u_char* data,
+void DHCP_Analyzer::DeliverPacket(uint64_t len, const u_char* data,
 			bool orig, uint64_t seq, const IP_Hdr* ip, int caplen)
 	{
 	Analyzer::DeliverPacket(len, data, orig, seq, ip, caplen);
