+++ src/analyzer/protocol/finger/Finger.h
@@ -15,7 +15,7 @@ public:
 
 	void Done() override;
 	// Line-based input.
-	void DeliverStream(int len, const u_char* data, bool orig) override;
+	void DeliverStream(uint64_t len, const u_char* data, bool orig) override;
 
 	static analyzer::Analyzer* Instantiate(Connection* conn)
 		{ return new Finger_Analyzer(conn); }
