+++ src/analyzer/protocol/gtpv1/GTPv1.h
@@ -11,7 +11,7 @@ public:
 	virtual ~GTPv1_Analyzer();
 
 	virtual void Done();
-	virtual void DeliverPacket(int len, const u_char* data, bool orig,
+	virtual void DeliverPacket(uint64_t len, const u_char* data, bool orig,
 					uint64_t seq, const IP_Hdr* ip, int caplen);
 
 	static analyzer::Analyzer* Instantiate(Connection* conn)
