+++ src/analyzer/protocol/http/HTTP.h
@@ -168,7 +168,7 @@ public:
 	// Overriden from Analyzer.
 	void Done() override;
 	void DeliverStream(uint64_t len, const u_char* data, bool orig) override;
-	void Undelivered(uint64_t seq, int len, bool orig) override;
+	void Undelivered(uint64_t seq, uint64_t len, bool orig) override;
 
 	// Overriden from tcp::TCP_ApplicationAnalyzer
 	void EndpointEOF(bool is_orig) override;
