+++ src/analyzer/protocol/krb/KRB.cc
@@ -64,7 +64,7 @@ void KRB_Analyzer::Done()
 	Analyzer::Done();
 	}
 
-void KRB_Analyzer::DeliverPacket(int len, const u_char* data, bool orig,
+void KRB_Analyzer::DeliverPacket(uint64_t len, const u_char* data, bool orig,
 				 uint64_t seq, const IP_Hdr* ip, int caplen)
 	{
 	Analyzer::DeliverPacket(len, data, orig, seq, ip, caplen);
