+++ src/analyzer/protocol/login/Login.h
@@ -26,7 +26,7 @@ public:
 	Login_Analyzer(const char* name, Connection* conn);
 	~Login_Analyzer() override;
 
-	void DeliverStream(int len, const u_char* data, bool orig) override;
+	void DeliverStream(uint64_t len, const u_char* data, bool orig) override;
 
 	void SetEnv(bool orig, char* name, char* val) override;
